+++ src/main/java/io/kontur/eventapi/dao/FeedDao.java
@@ -41,7 +41,8 @@ public class FeedDao {
         mapper.insertFeedData(feedData.getEventId(), feedData.getFeedId(), feedData.getVersion(),
                 feedData.getName(), feedData.getProperName(), feedData.getDescription(),
                 feedData.getStartedAt(), feedData.getEndedAt(), feedData.getUpdatedAt(),
-                feedData.getObservations(), episodesJson, feedData.getEnriched(), feedData.getUrls());
+                feedData.getObservations(), episodesJson, feedData.getEnriched(), feedData.getUrls(),
+                feedData.getLocation());
 
         mapper.markOutdatedEventsVersions(feedData.getEventId(), feedData.getFeedId(), feedData.getVersion());
         feedEventStatusDao.markAsActual(feedData.getFeedId(), feedData.getEventId(), true);