+++ src/main/java/io/kontur/eventapi/entity/KonturEvent.java
@@ -1,25 +1,16 @@
 package io.kontur.eventapi.entity;
 
+import lombok.Data;
+
 import java.util.*;
 
+@Data
 public class KonturEvent {
     private UUID eventId;
     private Set<UUID> observationIds = new HashSet<>();
 
     public KonturEvent(UUID eventId) {
         this.eventId = eventId;
-    }
-
-    public UUID getEventId() {
-        return eventId;
-    }
-
-    public void setEventId(UUID eventId) {
-        this.eventId = eventId;
-    }
-
-    public Set<UUID> getObservationIds() {
-        return observationIds;
     }
 
     public KonturEvent setObservationIds(Set<UUID> observationIds) {
@@ -29,13 +20,5 @@ public class KonturEvent {
 
     public void addObservations(UUID observations) {
         this.observationIds.add(observations);
-    }
-
-    @Override
-    public String toString() {
-        return "KonturEvent{" +
-                "eventId=" + eventId +
-                ", observationIds=" + observationIds +
-                '}';
     }
 }