+++ src/main/resources/db/changelog/v0.8/add-location-column-to-feed-data.sql
@@ -1,0 +1,7 @@
+--liquibase formatted sql
+
+--changeset event-api-migrations:v0.8/add-location-column-to-feed-data.sql runOnChange:false
+
+alter table feed_data add column if not exists location text;
+
+